+++ dotfiles/.config/ml4w/themes/glass/theme.sh
@@ -5,10 +5,6 @@
 echo "/ml4w-glass-center;/ml4w-glass-center/default" > $HOME/.config/ml4w/settings/waybar-theme.sh
 $HOME/.config/waybar/launch.sh &
 
-# Set nwg-dock-hyprland
-echo "glass" > $HOME/.config/ml4w/settings/dock-theme
-$HOME/.config/nwg-dock-hyprland/launch.sh &
-
 # Set swaync
 echo '@import "themes/glass/style.css";' > $HOME/.config/swaync/style.css
 swaync-client -rs