+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 2.15
 --------------------------------------------------------
+- nwg-dock-hyprland replaced with Quickshell based dock (nwg-dock-hyprland not supported on Fedora 44)
 - Scratchpad feature updated: SUPER + SHIFT + S to move the current window into the scratchpad workspace.
   SUPER + S to toggle the scatchpad workspace.
 - Fedora Copr updates